add SequenceSpaceJacobianType
