+++ examples/SequenceSpaceJacobian/estimation/us_data.csv
@@ -1,0 +1,157 @@
+t,pi,Y,i
+1966Q1,2.027726462326318,4409.518,4.56
+1966Q2,3.2809272256504096,4424.581,4.913333333333333
+1966Q3,3.3420896701509717,4462.053,5.41
+1966Q4,3.5106125343581374,4498.66,5.5633333333333335
+1967Q1,2.2172381579597777,4538.498,4.823333333333333
+1967Q2,2.7030988743893536,4541.28,3.9899999999999998
+1967Q3,3.6868999737552866,4584.246,3.893333333333333
+1967Q4,4.099099819771168,4618.812,4.173333333333333
+1968Q1,4.626964682421297,4713.013,4.79
+1968Q2,4.741267185358122,4791.758,5.983333333333333
+1968Q3,4.554924681617445,4828.892,5.946666666666666
+1968Q4,4.7283229880399125,4847.885,5.916666666666667
+1969Q1,4.4443613036874385,4923.76,6.566666666666666
+1969Q2,4.874927146116881,4938.728,8.326666666666666
+1969Q3,4.6977053489222875,4971.349,8.983333333333333
+1969Q4,4.779851747570585,4947.104,8.94
+1970Q1,4.434231410319089,4939.759,8.573333333333332
+1970Q2,4.588913164016928,4946.77,7.886666666666667
+1970Q3,4.5368729679635456,4992.357,6.706666666666667
+1970Q4,5.734082840626531,4938.857,5.566666666666666
+1971Q1,5.159697948356112,5072.996,3.856666666666667
+1971Q2,4.645064457771597,5100.447,4.566666666666666
+1971Q3,3.821648782308773,5142.422,5.476666666666667
+1971Q4,2.307614077587372,5154.547,4.75
+1972Q1,3.8868078576567155,5249.337,3.546666666666667
+1972Q2,2.7326716003723384,5368.485,4.3
+1972Q3,3.040745826667335,5419.184,4.743333333333333
+1972Q4,2.5347392821113868,5509.926,5.1466666666666665
+1973Q1,2.8178709096540855,5646.286,6.536666666666666
+1973Q2,5.557327598313422,5707.755,7.816666666666666
+1973Q3,5.430422724380013,5677.738,10.56
+1973Q4,5.675159521543982,5731.632,9.996666666666666
+1974Q1,7.048724156916686,5682.353,9.323333333333332
+1974Q2,10.320832607969322,5695.859,11.25
+1974Q3,11.93327124751411,5642.025,12.089999999999998
+1974Q4,10.136543231700212,5620.126,9.346666666666666
+1975Q1,8.052191302632838,5551.713,6.303333333333334
+1975Q2,6.227302116867195,5591.382,5.419999999999999
+1975Q3,6.146424601671652,5687.087,6.16
+1975Q4,6.671137814679029,5763.665,5.413333333333334
+1976Q1,6.222421474264173,5893.276,4.826666666666667
+1976Q2,4.877026223059522,5936.515,5.196666666666666
+1976Q3,6.419592932901952,5969.089,5.283333333333333
+1976Q4,6.40401493429672,6012.356,4.873333333333334
+1977Q1,6.899888264358478,6083.391,4.66
+1977Q2,6.262025373644642,6201.659,5.156666666666666
+1977Q3,6.820718838174811,6313.559,5.82
+1977Q4,5.939893600063284,6313.697,6.513333333333333
+1978Q1,6.386531545719865,6333.848,6.756666666666667
+1978Q2,7.08145796972961,6578.605,7.283333333333334
+1978Q3,7.039507513847321,6644.754,8.1
+1978Q4,7.381271218722252,6734.069,9.583333333333334
+1979Q1,5.576431308866536,6746.176,10.073333333333332
+1979Q2,9.084738020936655,6753.389,10.18
+1979Q3,7.592775224213666,6803.558,10.946666666666667
+1979Q4,8.61171263387459,6820.572,13.576666666666666
+1980Q1,10.146357608301116,6842.024,15.046666666666667
+1980Q2,9.25166483078057,6701.046,12.686666666666667
+1980Q3,9.208420477430735,6693.082,9.836666666666666
+1980Q4,10.108699875040884,6817.903,15.853333333333333
+1981Q1,9.170729490905384,6951.495,16.569999999999997
+1981Q2,7.778620671366254,6899.98,17.78
+1981Q3,7.572698513507259,6982.609,17.576666666666664
+1981Q4,7.3034302294325615,6906.529,13.586666666666666
+1982Q1,5.989278584097746,6799.233,14.226666666666667
+1982Q2,5.522610337039668,6830.251,14.513333333333334
+1982Q3,6.566642046124382,6804.139,11.006666666666666
+1982Q4,5.685935342708959,6806.857,9.286666666666667
+1983Q1,5.3622127483548665,6896.561,8.653333333333334
+1983Q2,3.0955934928774687,7053.5,8.803333333333335
+1983Q3,6.2110509440153505,7194.504,9.459999999999999
+1983Q4,3.240567944047057,7344.597,9.43
+1984Q1,4.106983472086934,7488.167,9.686666666666667
+1984Q2,4.759165604133675,7617.547,10.556666666666667
+1984Q3,3.8101479712298936,7690.985,11.39
+1984Q4,2.8770118253299115,7754.117,9.266666666666667
+1985Q1,5.5697204002868705,7829.26,8.476666666666667
+1985Q2,3.5157924557156273,7898.194,7.923333333333333
+1985Q3,4.172987950488971,8018.809,7.8999999999999995
+1985Q4,2.8976153297537444,8078.415,8.103333333333333
+1986Q1,4.394982546176385,8153.829,7.826666666666667
+1986Q2,2.7667184549602597,8190.552,6.919999999999999
+1986Q3,2.7771158166389975,8268.935,6.206666666666666
+1986Q4,3.2870213045291585,8313.338,6.266666666666667
+1987Q1,2.5377828702118466,8375.274,6.22
+1987Q2,3.934616254813439,8465.63,6.6499999999999995
+1987Q3,3.6258849448892017,8539.075,6.843333333333334
+1987Q4,4.12972938581988,8685.694,6.916666666666667
+1988Q1,3.915837635080588,8730.569,6.663333333333334
+1988Q2,4.938575999661254,8845.28,7.156666666666666
+1988Q3,4.686247770476903,8897.107,7.983333333333333
+1988Q4,4.545158407672822,9015.661,8.47
+1989Q1,4.437655204034674,9107.314,9.443333333333333
+1989Q2,3.687604365845787,9176.827,9.726666666666667
+1989Q3,3.010712647684022,9244.816,9.083333333333334
+1989Q4,3.711754627295849,9263.033,8.613333333333333
+1990Q1,4.687881581608444,9364.259,8.25
+1990Q2,4.5431131786016765,9398.243,8.243333333333332
+1990Q3,4.10518777470994,9404.494,8.16
+1990Q4,3.40114080232623,9318.876,7.743333333333332
+1991Q1,3.596553247938239,9275.276,6.426666666666667
+1991Q2,3.1186829191091725,9347.597,5.863333333333333
+1991Q3,3.651268658517526,9394.834,5.6433333333333335
+1991Q4,3.1836329258542184,9427.581,4.816666666666666
+1992Q1,3.09644363023045,9540.444,4.023333333333333
+1992Q2,2.8563887638782814,9643.893,3.7699999999999996
+1992Q3,2.390044216982856,9739.185,3.2566666666666664
+1992Q4,2.964125852679911,9840.753,3.0366666666666666
+1993Q1,2.783004173985537,9857.185,3.0399999999999996
+1993Q2,3.07071009671529,9914.565,3.0
+1993Q3,2.2556908169596435,9961.873,3.06
+1993Q4,2.0744170484439106,10097.362,2.99
+1994Q1,1.8284621988157612,10195.338,3.2133333333333334
+1994Q2,2.719689723611274,10333.495,3.94
+1994Q3,2.2529429759334585,10393.898,4.486666666666667
+1994Q4,2.0457251419062183,10512.962,5.166666666666667
+1995Q1,2.112723917077197,10550.251,5.81
+1995Q2,2.395348900058236,10581.723,6.02
+1995Q3,1.9233521188806169,10671.738,5.796666666666667
+1995Q4,2.0210880472447235,10744.203,5.72
+1996Q1,1.5786404616660477,10824.674,5.363333333333333
+1996Q2,1.9167858345777278,11005.217,5.243333333333333
+1996Q3,1.8379638284230282,11103.935,5.306666666666667
+1996Q4,2.2961949165604123,11219.238,5.28
+1997Q1,1.5806884724582915,11291.665,5.276666666666666
+1997Q2,2.1028090646867836,11479.33,5.523333333333333
+1997Q3,0.9365140059810262,11622.911,5.533333333333334
+1997Q4,1.2733258670848402,11722.722,5.506666666666667
+1998Q1,1.1825584715647208,11839.876,5.52
+1998Q2,1.230948099766982,11949.492,5.5
+1998Q3,1.4290518121450324,12099.191,5.533333333333334
+1998Q4,1.277794248941766,12294.737,4.86
+1999Q1,1.0101467687091858,12410.778,4.733333333333333
+1999Q2,1.3953198458593885,12514.408,4.746666666666667
+1999Q3,1.480837465298146,12679.977,5.093333333333334
+1999Q4,1.8940495006775304,12888.281,5.306666666666667
+2000Q1,2.1126402624124463,12935.252,5.676666666666667
+2000Q2,1.4035489543970048,13170.749,6.273333333333333
+2000Q3,1.9409855161090617,13183.89,6.52
+2000Q4,1.9985997069409622,13262.25,6.473333333333334
+2001Q1,2.6251602936705476,13219.251,5.593333333333334
+2001Q2,1.5195002008354441,13301.394,4.326666666666667
+2001Q3,1.0378460852063842,13248.142,3.4966666666666666
+2001Q4,1.943387349712089,13284.881,2.1333333333333333
+2002Q1,1.23520367779284,13394.91,1.7333333333333334
+2002Q2,2.2204300258506393,13477.356,1.75
+2002Q3,2.14639049932428,13531.741,1.74
+2002Q4,1.4302514400940325,13549.421,1.4433333333333334
+2003Q1,1.2409818612027612,13619.434,1.25
+2003Q2,1.5262288140348357,13741.107,1.2466666666666668
+2003Q3,1.8425461114967678,13970.157,1.0166666666666666
+2003Q4,1.803906908203734,14131.379,0.9966666666666667
+2004Q1,2.2187632117856237,14212.34,1.0033333333333332
+2004Q2,2.3043794032079568,14323.017,1.01
+2004Q3,1.4414629413773028,14457.832,1.4333333333333333
+2004Q4,2.1859911994692727,14605.595,1.95
